Thêm reservation status entity, service, impl và xóa code dư

+++ src/main/java/com/springleaf_restaurant_backend/user/entities/ReservationStatus.java
@@ -13,7 +13,7 @@ public class ReservationStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reservation_status_id")
-    private Long reservationId;
+    private Integer reservationId;
 
     @Column(name = "reservation_status_name")
     private String restaurantTableId;
